+++ src/components/Payment/PaymentListItem.tsx
@@ -1,4 +1,4 @@
-import formatCurrency from '@/utils/useCurrencyFormat';
+import formatCurrency from '@/utils/currencyFormat';
 
 interface PaymentListItemProp {
   price: number;
@@ -11,7 +11,7 @@ const PaymentListItem = ({
 }: PaymentListItemProp): JSX.Element => {
   // const formatCurrency = formatCurrency();
   return (
-    <li className=" border-b border-[#CAD4E7] px-[0] py-[10px] flex justify-between items-start  text-[#000] text-sm not-italic font-light leading-[24px]">
+    <li className="flex items-start justify-between border-b border-[#CAD4E7] px-[0] py-[0.625rem] text-sm font-light not-italic leading-[1.5rem] text-[#000]">
       <span>{children}</span>
       <span>{formatCurrency(price)} 원</span>
     </li>
